fix(weapons): cast rarity into short

+++ src/entities/weapon/weapon.c
@@ -13,7 +13,7 @@ void *create_weapon_from_db(sqlite3_stmt *stmt) {
     weapon->min_attack = sqlite3_column_int(stmt, 2);
     weapon->max_attack = sqlite3_column_int(stmt, 3);
     weapon->attacks_per_turn = sqlite3_column_int(stmt, 4);
-    weapon->rarity = sqlite3_column_int(stmt, 5);
+    weapon->rarity = (short)sqlite3_column_int(stmt, 5);
     weapon->max_uses = sqlite3_column_int(stmt, 6);
     weapon->uses = weapon->max_uses;
     weapon->cost = sqlite3_column_int(stmt, 7);
